date changed to string and type resources created
