+++ source/SIL.AppBuilder.Portal.Frontend/src/data/index.ts
@@ -1,4 +1,5 @@
 export { default as DataProvider } from './provider';
+import { USERS_TYPE, REVIEWERS_TYPE, PRODUCTS_TYPE, GROUP_MEMBERSHIPS_TYPE, ORGANIZATION_MEMBERSHIPS_TYPE } from '@data';
 
 export { defaultSourceOptions, defaultOptions } from './store';
 export {
@@ -32,14 +33,11 @@ export { pushPayload } from './push-payload';
 
 export { TASKS_TYPE, TaskResource } from './models/task';
 export { NOTIFICATIONS_TYPE, NotificationResource } from './models/notification';
-
-// JSONAPI types
-export type ORGANIZATIONS_TYPE = 'organizations';
-export type GROUPS_TYPE = 'groups';
-export type PROJECTS_TYPE = 'projects';
-export type USERS_TYPE = 'users';
-export type PRODUCTS_TYPE = 'products';
-export type ORGANIZATION_MEMBERSHIPS_TYPE = 'organization-memberships';
-export type GROUP_MEMBERSHIPS_TYPE = 'group-memberships';
-
-export type REVIEWERS_TYPE = 'reviewers';
+export { ORGANIZATIONS_TYPE, OrganizationResource } from './models/organization';
+export { GROUPS_TYPE, GroupResource } from './models/group';
+export { PROJECTS_TYPE, ProjectResource } from './models/project';
+export { USERS_TYPE, UserResource } from './models/user';
+export { REVIEWERS_TYPE, ReviewerResource } from './models/reviewer';
+export { PRODUCTS_TYPE, ProductResource } from './models/product';
+export { GROUP_MEMBERSHIPS_TYPE, GroupMembershipResource } from './models/group-membership';
+export { ORGANIZATION_MEMBERSHIPS_TYPE, OrganizationMembershipResource } from './models/organization-membership';
