+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/group-membership.ts
@@ -1,7 +1,11 @@
-import { AttributesObject } from "jsonapi-typescript";
+import { AttributesObject, ResourceObject } from "jsonapi-typescript";
+
+export type GROUP_MEMBERSHIPS_TYPE = 'group-memberships';
 
 export const TYPE_NAME = 'group-membership';
 export const PLURAL_NAME = 'group-memberships';
 
 export interface GroupMembershipAttributes extends AttributesObject {
 }
+
+export type GroupMembershipResource = ResourceObject<GROUP_MEMBERSHIPS_TYPE, GroupMembershipAttributes>;
