+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/group.ts
@@ -1,4 +1,6 @@
-import { AttributesObject } from "jsonapi-typescript";
+import { AttributesObject, ResourceObject } from "jsonapi-typescript";
+
+export type GROUPS_TYPE = 'groups';
 
 export const TYPE_NAME = 'group';
 export const PLURAL_NAME = 'groups';
@@ -7,3 +9,5 @@ export interface GroupAttributes extends AttributesObject {
   name: string;
   abbreviation: string;
 }
+
+export type GroupResource = ResourceObject<GROUPS_TYPE, GroupAttributes>;
