+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/notification.ts
@@ -7,7 +7,7 @@ export const TYPE_NAME = 'notification';
 export interface NotificationAttributes extends AttributesObject {
   title: string;
   description: string;
-  time: Date;
+  time: string;
   link: string;
   isViewed: boolean;
   show: boolean;
