+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/organization.ts
@@ -1,4 +1,6 @@
-import { AttributesObject } from "jsonapi-typescript";
+import { AttributesObject, ResourceObject } from "jsonapi-typescript";
+
+export type ORGANIZATIONS_TYPE = 'organizations';
 
 export const TYPE_NAME = 'organization';
 
@@ -22,3 +24,5 @@ export interface OrganizationAttributes extends AttributesObject {
   // TODO: find out if we can define custom getters and setters on 'Models'
   logoUrl?: string;
 }
+
+export type OrganizationResource = ResourceObject<ORGANIZATIONS_TYPE, OrganizationAttributes>;
