+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/product.ts
@@ -1,4 +1,7 @@
-import { AttributesObject } from "jsonapi-typescript";
+import { AttributesObject, ResourceObject } from "jsonapi-typescript";
+
+export type PRODUCTS_TYPE = 'products';
+
 
 export const TYPE_NAME = 'product';
 export const PLURAL_NAME = 'products';
@@ -9,3 +12,5 @@ export interface ProductAttributes extends AttributesObject {
   // TODO: figure out better mapping for this
   //       we'll know for certain as we actually start to work with products
 }
+
+export type ProductResource = ResourceObject<PRODUCTS_TYPE, ProductAttributes>;
