+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/reviewer.ts
@@ -1,4 +1,6 @@
-import { AttributesObject } from "jsonapi-typescript";
+import { AttributesObject, ResourceObject } from "jsonapi-typescript";
+
+export type REVIEWERS_TYPE = 'reviewers';
 
 export const TYPE_NAME = 'reviewer';
 export const PLURAL_NAME = 'reviewers';
@@ -7,3 +9,5 @@ export interface ReviewerAttributes extends AttributesObject {
   name: string;
   email: string;
 }
+
+export type ReviewerResource = ResourceObject<REVIEWERS_TYPE, ReviewerAttributes>;
