+++ source/SIL.AppBuilder.Portal.Frontend/src/data/models/user.ts
@@ -1,4 +1,6 @@
-import { AttributesObject } from 'jsonapi-typescript';
+import { AttributesObject, ResourceObject } from 'jsonapi-typescript';
+
+export type USERS_TYPE = 'users';
 
 export const TYPE_NAME = 'user';
 export const PLURAL_NAME = 'users';
@@ -24,3 +26,5 @@ export interface UserAttributes extends AttributesObject {
 export function name(attrs: UserAttributes) {
   return `${attrs.givenName} ${attrs.familyName}`;
 }
+
+export type UserResource = ResourceObject<USERS_TYPE, UserAttributes>;
